pyproject.toml
[project]
name = "projeto-lakehouse"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apache-airflow==3.0.2",
    "apache-airflow-providers-apache-spark>=5.3.0",
    "apache-airflow-providers-papermill>=3.11.0",
    "ipykernel>=6.29.5",
]
